Final_Project/requirements.txt
torch --index-url https://download.pytorch.org/whl/cu113
torchvision --index-url https://download.pytorch.org/whl/cu113
torchaudio --index-url https://download.pytorch.org/whl/cu113
gdown
numpy
pandas
matplotlib
tqdm
scipy
pillow
ml-collections
pyyaml
logger
opencv-python